Merged PR 290: Fixed the drasi namespace set bug
I previous implemented it in a way so that you need to specify the namespace as a flag instead of an argument (e.g. drasi namespace set -n demo)

This is very confusing and I have changed so that the namespace will the inputted as an argument instead (e.g. drasi namespace set demo)

Related work items: #10656